Stepwise detection of recombination breakpoints using MaxChi

Stepwise detection of recombination breakpoints using the maximum chi-square (MaxChi) method at each step

maxchi(input_file, breaks, winHalfWidth, permReps)

Arguments

input_file
character string indicating the name of a Phylip format data input file

breaks
an integer vector of ordered site(s) just before the previously declared breakpoints

winHalfWidth
the window half width to use

permReps
the number of Monte Carlo replicates to use for the permutation distribution

Details

The maxchi function implements the maximum chi-square (MaxChi) method for detecting recombination breakpoints (Maynard Smith 1992) using a moving window of fixed width. Breakpoints detected in previous steps of a stepwise search may be conditioned upon.

For a given position of the moving window on the sequence alignment, and for a given pair of sequences, a chi-square statistic is computed to compare two proportions: the proportion of sites at which the sequences agree in the left half-window and the proportion of sites in at which the sequences agree in the right half-window. Discordance between the two proportions may reflect a recombination event, located at the window centre, in the history of the two sequences. The maximum chi-square over all sequence pairs is regarded as a summary of the evidence for recombination at the window centre. The individual chi-square statistics may also be of interest for suggesting pairs of sequences segments that descend from historical recombination events. Significance of observed chi-square statistics is assessed by a Monte Carlo permutation test. When conditioning on breakpoints proposed at previous steps of a stepwise search, permutation is restricted to sites within blocks defined by the previously proposed breakpoints, as described by Graham et al. (2004).

Value

polyposn

The site numbers of all ungapped polymorphic sites in the alignment

chisqs

Observed chi-square statistics that exceed the 90th percentile of the permutation null distribution

winlocs

Window centres corresponding to the chi-square statistics in chisqs

pairmem1

First member of each pair that lead to a significant chi-square statistic in chisqs

pairmem2

Second member of each pair

quants

90th and 95th percentiles of the permutation distribution

References

Graham J, McNeney B and Seillier-Moiseiwitsch F (2004). Stepwise detection of recombination breakpoints in sequence alignments. Bioinformatics Sep 23; [Epub ahead of print]

Maynard Smith J (1992). Analyzing the mosaic structure of genes. J Mol Evol, 34:126-129.

http://stat-db.stat.sfu.ca/stepwise

See also

Examples

 demo(maxchi)

#> 
#> 
#> 	demo(maxchi)
#> 	---- ~~~~~~
#> 
#> > inputFile<- system.file("extdata", "simulinfile.txt", package = "stepwise")
#> 
#> > breaks<-c(531,616,810)
#> 
#> > WinHalfWidth<-30
#> 
#> > permReps<-10
#> 
#> > a<-maxchi(inputFile, breaks, WinHalfWidth, permReps)
#> 
#> > summary.maxchi(a)
#> --------------------------------------------------
#> There were 15 site-specific MaxChi statistics significant at the
#> 10 percent level (90th percentile = 18.468, 95th percentile = 18.468):
#> 
#> Number Location  MaxChi   pairs
#>      1      531  20.000*   (sample17__:sample27__)
#>      2      548  20.000*   (sample17__:sample27__)
#>      3      549  20.000*   (sample17__:sample27__)
#>      4      556  20.000*   (sample17__:sample27__)
#>      5      616  20.000*   (sample19__:sample30__)
#>      6      807  19.461*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#>      7      810  24.310*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#>      8      811  19.461*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#> 			   (sample3___:sample11__)
#> 			   (sample3___:sample15__)
#> 			   (sample3___:sample16__)
#>      9      815  19.461*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#> 			   (sample3___:sample11__)
#> 			   (sample3___:sample15__)
#> 			   (sample3___:sample16__)
#>     10      826  19.200*   (sample3___:sample14__)
#> 			   (sample3___:sample27__)
#>     11      827  21.172*   (sample3___:sample14__)
#> 			   (sample3___:sample27__)
#>     12      828  19.200*   (sample4___:sample14__)
#> 			   (sample4___:sample27__)
#>     13      832  21.172*   (sample4___:sample14__)
#> 			   (sample4___:sample27__)
#>     14      834  21.172*   (sample4___:sample13__)
#> 			   (sample4___:sample14__)
#> 			   (sample4___:sample27__)
#>     15      844  21.172*   (sample4___:sample13__)
#> 			   (sample4___:sample14__)
#> 			   (sample4___:sample27__)
#> --------------------------------------------------
#> Notes - "Location" is the polymorphic site just before the proposed breakpoint.
#>       - MaxChi statistics significant at the 5 percent level indicated by a * 
#>