Stepwise detection of recombination breakpoints using Phylpro

Stepwise detection of recombination breakpoints using phylogenetic profiling (Phylpro) at each step

phylpro(input_file, breaks, winHalfWidth, permReps)

Arguments

input_file
character string indicating the name of a Phylip format data input file

breaks
an integer vector of ordered site(s) just before the previously declared breakpoints

winHalfWidth
the window half width to use

permReps
the number of Monte Carlo replicates to use for the permutation distribution

Details

The phylpro function implements phylogenetic profiling (Phylpro) for detecting recombination breakpoints (Weiller 1998) using a moving window of fixed width. Breakpoints detected in previous steps of a stepwise search may be conditioned upon.

For a given position of the moving window on the sequence alignment, and for a given “target” sequence, a correlation is computed to compare two distance vectors: the distance between the target sequence and all other sequences in the left half-window and the distance between the target sequence and all others in the right half-window. The pair-wise distance measure used is the proportion of sites at which the sequences differ. Discordance between the two distance vectors may reflect a recombination event, located at the window centre, in the history of the target sequence. The minimum correlation over all target sequences is regarded as a summary of the evidence for recombination at the window centre. The individual correlations for the target sequences may also be of interest for suggesting sequence segments that descend from historical recombination events. Significance of observed correlation statistics is assessed by a Monte Carlo permutation test. When conditioning on breakpoints proposed at previous steps of a stepwise search, permutation is restricted to sites within blocks defined by the previously proposed breakpoints, as described by Graham et al. (2004).

Value

polyposn

The site numbers of all ungapped polymorphic sites in the alignment

corrs

Observed correlations that exceed the 90th percentile of the permutation null distribution

winlocs

Window centres corresponding to the correlations in corrs

target.seqs

The target sequence that lead to a significant correlation in corrs

quants

90th and 95th percentiles of the permutation distribution

References

Graham J, McNeney B and Seillier-Moiseiwitsch F (2004). Stepwise detection of recombination breakpoints in sequence alignments. Bioinformatics Sep 23; [Epub ahead of print]

Weiller G (1998). Phylogenetic profiles: A graphical method for detecting genetic recombination in homologous sequences. Mol Biol Evol, 15:326-335.

http://stat-db.stat.sfu.ca/stepwise

See also

Examples

 demo(phylpro)

#> 
#> 
#> 	demo(phylpro)
#> 	---- ~~~~~~~
#> 
#> > inputFile<-system.file("extdata", "simulinfile.txt", package = "stepwise")
#> 
#> > breaks<-c(548, 735, 832)
#> 
#> > WinHalfWidth<-30
#> 
#> > permReps<-10
#> 
#> > b<-phylpro(inputFile, breaks, WinHalfWidth, permReps)
#> 
#> > summary.phylpro(b)
#> --------------------------------------------------
#> There were 20 site-specific Minimum Correlation statistics significant at the
#> 10 percent level (10th percentile = -0.460, 5th percentile = -0.460):
#> 
#> Number Location  MinCor   targets
#>      1      513  -0.549*   sample28__
#>      2      526  -0.656*   sample28__
#>      3      527  -0.627*   sample28__
#>      4      528  -0.583*   sample28__
#>      5      531  -0.648*   sample28__
#>      6      548  -0.696*   sample28__
#>      7      549  -0.691*   sample28__
#>      8      556  -0.684*   sample28__
#>      9      558  -0.635*   sample28__
#>     10      564  -0.631*   sample28__
#>     11      568  -0.625*   sample28__
#>     12      573  -0.641*   sample28__
#>     13      576  -0.528*   sample28__
#>     14      579  -0.539*   sample28__
#>     15      583  -0.536*   sample28__
#>     16      588  -0.616*   sample28__
#>     17      604  -0.509*   sample28__
#>     18      616  -0.475*   sample28__
#>     19      735  -0.492*   sample29__
#>     20      744  -0.460*   sample29__
#> --------------------------------------------------
#> Notes - "Location" is the polymorphic site just before the proposed breakpoint.
#>       - MinCor statistics significant at the 5 percent level indicated by a * 
#>