Summary of a maxchi object

Print a summary of a maxchi object

# S3 method for maxchi
summary(object, ...)

Arguments

object
A list object of class maxchi output by the maxchi function

...
Additional arguments to the summary function (currently unused)

Details

The output of maxchi includes information on all site- and pair-specific chi-squares that exceed the 90th percentile of the permutation distribution. The summary.maxchi function computes maximum site-specific measures (i.e. maxima over all sequence-pair chi-squares from the same site) and reports all sequence pairs that tied for these maximum values.

Value

siteWinlocs

Window locations with significant recombination signal

siteChisqs

The maximum chi-square at these window locations

pairs

The pair(s) that give the maximum chi-square statistic at each window location

See also

Examples

 demo(maxchi)

#> 
#> 
#> 	demo(maxchi)
#> 	---- ~~~~~~
#> 
#> > inputFile<- system.file("extdata", "simulinfile.txt", package = "stepwise")
#> 
#> > breaks<-c(531,616,810)
#> 
#> > WinHalfWidth<-30
#> 
#> > permReps<-10
#> 
#> > a<-maxchi(inputFile, breaks, WinHalfWidth, permReps)
#> 
#> > summary.maxchi(a)
#> --------------------------------------------------
#> There were 43 site-specific MaxChi statistics significant at the
#> 10 percent level (90th percentile = 16.484, 95th percentile = 16.484):
#> 
#> Number Location  MaxChi   pairs
#>      1      342  17.143*   (sample15__:sample28__)
#> 			   (sample16__:sample28__)
#>      2      453  16.596*   (sample20__:sample24__)
#>      3      463  18.261*   (sample20__:sample24__)
#>      4      470  18.261*   (sample20__:sample24__)
#>      5      528  16.705*   (sample17__:sample27__)
#>      6      531  20.000*   (sample17__:sample27__)
#>      7      548  20.000*   (sample17__:sample27__)
#>      8      549  20.000*   (sample17__:sample27__)
#>      9      556  20.000*   (sample17__:sample27__)
#>     10      558  18.261*   (sample17__:sample27__)
#>     11      564  18.261*   (sample17__:sample27__)
#>     12      568  18.261*   (sample17__:sample27__)
#>     13      573  18.261*   (sample17__:sample27__)
#>     14      576  16.596*   (sample17__:sample27__)
#>     15      588  18.373*   (sample17__:sample20__)
#>     16      604  18.373*   (sample17__:sample20__)
#>     17      616  20.000*   (sample19__:sample30__)
#>     18      618  16.705*   (sample19__:sample30__)
#>     19      620  18.468*   (sample19__:sample30__)
#>     20      622  18.468*   (sample19__:sample30__)
#>     21      633  17.330*   (sample21__:sample30__)
#>     22      637  17.330*   (sample21__:sample30__)
#>     23      644  17.330*   (sample21__:sample30__)
#>     24      649  17.330*   (sample21__:sample30__)
#>     25      752  17.143*   (sample29__:sample30__)
#>     26      753  17.143*   (sample29__:sample30__)
#>     27      756  17.143*   (sample29__:sample30__)
#>     28      805  17.376*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#>     29      807  19.461*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#>     30      810  24.310*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#>     31      811  19.461*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#> 			   (sample3___:sample11__)
#> 			   (sample3___:sample15__)
#> 			   (sample3___:sample16__)
#>     32      815  19.461*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#> 			   (sample3___:sample11__)
#> 			   (sample3___:sample15__)
#> 			   (sample3___:sample16__)
#>     33      818  17.143*   (sample3___:sample8___)
#> 			   (sample3___:sample9___)
#> 			   (sample3___:sample11__)
#> 			   (sample3___:sample15__)
#> 			   (sample3___:sample16__)
#>     34      819  17.330*   (sample3___:sample14__)
#> 			   (sample3___:sample27__)
#>     35      825  17.330*   (sample3___:sample14__)
#> 			   (sample3___:sample27__)
#>     36      826  19.200*   (sample3___:sample14__)
#> 			   (sample3___:sample27__)
#>     37      827  21.172*   (sample3___:sample14__)
#> 			   (sample3___:sample27__)
#>     38      828  19.200*   (sample4___:sample14__)
#> 			   (sample4___:sample27__)
#>     39      832  21.172*   (sample4___:sample14__)
#> 			   (sample4___:sample27__)
#>     40      834  21.172*   (sample4___:sample13__)
#> 			   (sample4___:sample14__)
#> 			   (sample4___:sample27__)
#>     41      844  21.172*   (sample4___:sample13__)
#> 			   (sample4___:sample14__)
#> 			   (sample4___:sample27__)
#>     42      854  18.373*   (sample13__:sample18__)
#> 			   (sample14__:sample18__)
#> 			   (sample18__:sample27__)
#>     43      858  18.373*   (sample1___:sample13__)
#> 			   (sample1___:sample14__)
#> 			   (sample1___:sample27__)
#> 			   (sample4___:sample13__)
#> 			   (sample4___:sample14__)
#> 			   (sample4___:sample27__)
#> 			   (sample5___:sample13__)
#> 			   (sample5___:sample14__)
#> 			   (sample5___:sample27__)
#> 			   (sample6___:sample13__)
#> 			   (sample6___:sample14__)
#> 			   (sample6___:sample27__)
#> 			   (sample7___:sample13__)
#> 			   (sample7___:sample14__)
#> 			   (sample7___:sample27__)
#> --------------------------------------------------
#> Notes - "Location" is the polymorphic site just before the proposed breakpoint.
#>       - MaxChi statistics significant at the 5 percent level indicated by a * 
#>