Summary of a phylpro object

Print a summary of a phylpro object

# S3 method for phylpro
summary(object, ...)

Arguments

object
A list object of class phylpro output by the phylpro function

...
Additional arguments to the summary function (currently unused)

Details

The output of phylpro includes information on all site- and target sequence-specific correlations that exceed the 90th percentile of the permutation distribution. The summary.phylpro function computes minimum site-specific measures (i.e. minima over all target-sequence correlations from the same site) and reports all target sequences that tied for these minimum values.

Value

siteWinlocs

Window locations with significant recombination signal

siteCorrs

The minimum correlation at these window locations

target.seqs

The target sequence(s) that give the minimum correlation at each window location

See also

Examples

 demo(phylpro)

#> 
#> 
#> 	demo(phylpro)
#> 	---- ~~~~~~~
#> 
#> > inputFile<-system.file("extdata", "simulinfile.txt", package = "stepwise")
#> 
#> > breaks<-c(548, 735, 832)
#> 
#> > WinHalfWidth<-30
#> 
#> > permReps<-10
#> 
#> > b<-phylpro(inputFile, breaks, WinHalfWidth, permReps)
#> 
#> > summary.phylpro(b)
#> --------------------------------------------------
#> There were 38 site-specific Minimum Correlation statistics significant at the
#> 10 percent level (10th percentile = -0.251, 5th percentile = -0.251):
#> 
#> Number Location  MinCor   targets
#>      1      437  -0.330*   sample16__
#>      2      438  -0.282*   sample16__
#>      3      463  -0.391*   sample28__
#>      4      470  -0.384*   sample28__
#>      5      483  -0.314*   sample28__
#>      6      486  -0.295*   sample28__
#>      7      496  -0.442*   sample28__
#>      8      513  -0.549*   sample28__
#>      9      526  -0.656*   sample28__
#>     10      527  -0.627*   sample28__
#>     11      528  -0.583*   sample28__
#>     12      531  -0.648*   sample28__
#>     13      548  -0.696*   sample28__
#>     14      549  -0.691*   sample28__
#>     15      556  -0.684*   sample28__
#>     16      558  -0.635*   sample28__
#>     17      564  -0.631*   sample28__
#>     18      568  -0.625*   sample28__
#>     19      573  -0.641*   sample28__
#>     20      576  -0.528*   sample28__
#>     21      579  -0.539*   sample28__
#>     22      583  -0.536*   sample28__
#>     23      588  -0.616*   sample28__
#>     24      604  -0.509*   sample28__
#>     25      616  -0.475*   sample28__
#>     26      618  -0.415*   sample28__
#>     27      620  -0.342*   sample28__
#>     28      622  -0.356*   sample28__
#>     29      735  -0.492*   sample29__
#>     30      737  -0.406*   sample29__
#>     31      741  -0.449*   sample29__
#>     32      742  -0.381*   sample29__
#>     33      744  -0.460*   sample29__
#>     34      746  -0.345*   sample29__
#>     35      747  -0.336*   sample29__
#>     36      750  -0.323*   sample29__
#>     37      752  -0.329*   sample29__
#>     38      756  -0.342*   sample29__
#> --------------------------------------------------
#> Notes - "Location" is the polymorphic site just before the proposed breakpoint.
#>       - MinCor statistics significant at the 5 percent level indicated by a * 
#>